Struct xori::arch::x86::analyzex86::Analysisx86

pub struct Analysisx86 {
    pub xi: Xori,
    pub header: Header,
    pub base: usize,
    pub address_tracker: BTreeMap<u64, usize>,
    pub instr_info: BTreeMap<u64, InstructionInfo>,
    pub functions: Vec<FuncInfo>,
    pub symbols: Option<Vec<Symbol>>,
    pub sig_analyzer: SigAnalyzer,
}

Fields
xi: Xori header: Header base: usize address_tracker: BTreeMap<u64, usize> instr_info: BTreeMap<u64, InstructionInfo> functions: Vec<FuncInfo> symbols: Option<Vec<Symbol>> sig_analyzer: SigAnalyzer
Methods

impl Analysisx86

pub fn add_jump(&mut self, address: &u64, offset: u64, left: i64, right: i64)

Add the jump address to the state's current function jump table.

pub fn add_return(&mut self, address: &u64, offset: u64)

Add the address of the return to the state's current function return list.

pub fn add_return_value(
&mut self,
value: u64,
address: &u64,
offset: u64
) -> (bool, u64)

Check if the return value is valid

pub fn add_func(
&mut self,
calling_function: u64,
offset: u64,
return_offset: u64,
memory_address: u64,
destination_offset: u64,
memory_type: MemoryType,
import_only: bool
) -> Option<String>

Add the dest address of the call to the function tracker if the dest address already exists add the xref address to the list Check if the dest address points to an import apicall